corret the enum to its respactive folder and remove prefix user- from pusher client subscripition
