+++ src/services/notify/types/events.ts
@@ -1,0 +1,9 @@
+export enum NotificationEvent {
+  CHAT_INVITE = "chat_invite",
+  NEW_CONNECTION = "new_connection",
+  CONNECTION_ACCEPTED = "accept_connection",
+  UPDATE_TASK = "update_task",
+  TASK_ASSIGNED = "task_assigned",
+  TASK_COMMENTED = "task_commented",
+  PROJECT_INVITE = "project_invite"
+}
